+++ src/map/IMapInterfaces.ts
@@ -35,7 +35,7 @@ export interface IMapCore {
 export interface IMapZoomController {
     // Method the UI calls to change the map zoom
     setZoom(zoomLevel: number): void;
-    // Method the Map Adapter calls to update the central store
+    // Method the Map Adapter calls to update the map state store
     notifyZoomChange(zoomLevel: number): void;
 }